requirements.txt
# torch
# numpy
# torchvision
# transformers
# accelerate
# ftfy
# git+https://github.com/huggingface/diffusers
# modin[all]
Pillow
openxlab-test